My first attempt at cleaning up the frontend js code.
I think all classes should be instantiated in app.ts and use dependency
injection to construct a single DB and socket instance and inject them
to those who need them:
http://misko.hevery.com/2008/09/10/where-have-all-the-new-operators-gone/
Then may be we can solve the problem to the DB and socket been
constructed twice.

We start first by fetching all data needed by all pages (but just index
page for now) in Socket.ts and let the individual pages retrieve this
from the global windows.appInterface if the data is there i.e fetched
successfully while online or from the database (we are offline or
liefi).